Added password verified the product page
