--- src/test/java/LaunchBrowser.java
@@ -1,20 +1,0 @@
-import com.microsoft.playwright.*;
-import org.testng.annotations.Test;
-
-public class LaunchBrowser {
-
-    @Test
-    public void launchChromium() {
-        try (Playwright playwright = Playwright.create()) {
-            Browser browser = playwright.chromium().launch(new BrowserType.LaunchOptions().setHeadless(false));
-            BrowserContext context = browser.newContext();
-            Page page = context.newPage();
-            page.navigate("https://www.saucedemo.com");
-            page.fill("#user-name", "standard_user");
-            Thread.sleep(1000);
-
-        } catch (InterruptedException e) {
-            throw new RuntimeException(e);
-        }
-    }
-}
